User update changed

+++ backend/demo/src/main/java/com/ChitChat/demo/business/concretes/UserManager.java
@@ -6,6 +6,8 @@ import com.ChitChat.demo.dto.requests.UserUpdateRequest;
 import com.ChitChat.demo.dto.responses.UserVM;
 import com.ChitChat.demo.entity.User;
 import com.ChitChat.demo.error.PasswordMismatchException;
+import com.ChitChat.demo.error.SamePasswordException;
+import com.ChitChat.demo.error.UsernameAlreadyExistsException;
 import com.ChitChat.demo.repository.TokenRepository;
 import com.ChitChat.demo.repository.UserRepository;
 import com.ChitChat.demo.mapper.ModelMapperService;
@@ -15,6 +17,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -35,6 +38,11 @@ public class UserManager implements UserService {
     private PasswordEncoder passwordEncoder;
     @Override
     public void save(UserRegisterRequest userRegisterRequest) {
+
+        if(userRepository.existsByUsername(userRegisterRequest.getUsername())){
+            throw new UsernameAlreadyExistsException();
+        }
+
         String encodedPassword = passwordEncoder.encode(userRegisterRequest.getPassword());
         User user = new User();
         user.setUsername(userRegisterRequest.getUsername());
@@ -76,12 +84,22 @@ public class UserManager implements UserService {
                 throw new PasswordMismatchException();
             }
             //if matches
+
+            //if the new password is the same as the previous password
+            if(passwordEncoder.matches(userUpdateRequest.getNewPassword(), user.getPassword())){
+                throw new SamePasswordException();
+            }
+
+
             tokenRepository.deleteByUserId(id);
             user.setPassword(passwordEncoder.encode(userUpdateRequest.getNewPassword()));
 
         }
 
         if(userUpdateRequest.getUsername() != null){
+            if(userRepository.existsByUsername(userUpdateRequest.getUsername())){
+                throw new UsernameAlreadyExistsException();
+            }
             user.setUsername(userUpdateRequest.getUsername());
         }
